Menu do jogo melhorado
