+++ src/components/carteiro.rs
@@ -17,9 +17,10 @@ impl Carteiro{
 
 
     Carteiro {
-        pos_x: 1,
-        pos_y: 1,
+        pos_x: 0,
+        pos_y: 0,
         status: Status::Menu
     }
     }
+
 }
